[class] feat : 커리큘럼 레슨 페이지 리팩토링
